@@ -10,8 +10,8 @@
     "mobx": "^6.3.10",
     "mobx-react-lite": "^3.2.3",
     "pluralize": "^8.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-dragula": "^1.1.17",
     "react-scripts": "^5.0.0",
     "uuid": "^8.3.2",
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.2.1",
     "codecov": "^3.8.3",
     "mini-css-extract-plugin": "2.4.5",
